+++ source.q
@@ -1,0 +1,2 @@
+let matrix = [1, 2, 3|
+             |1, 2, 4]a;
