+++ src/compiler/compiler.rs
@@ -29,7 +29,8 @@ impl Compile for String
 			.map(|line| format!("{line}\n").chars().collect())
 			.collect();
 
-		let target = dbg!(self.lex(&source)?)
+		let target = self
+			.lex(&source)?
 			.parse(&source)?
 			.analyse(&source)?
 			.synthesise();
