+++ src/compiler/parser/rules/mod.rs
@@ -2,5 +2,6 @@ mod declaration;
 mod expression;
 mod programme;
 mod statement;
+mod utils;
 
 pub use super::*;
